Moved app theme configuration into separate file
