+++ lib/theme/app_theme.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+import 'package:layout_app/theme/custom_colors.dart';
+
+final ThemeData appTheme = ThemeData(
+  primaryColor: Colors.lightBlue,
+  primarySwatch: CustomColors.onlyWhite,
+);
